Consolidate date/time utils in dateTime.ts
