+++ src/components/dataPanel/FieldValue.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import { Placeholders } from "../../utils/mcaWritingStyleFormatter";
+import { Placeholders } from "../../utils/writingStyle";
 import { FieldValue } from "./FieldValue";
 
 describe("FieldValue", () => {
