+++ src/components/dataPanel/PanelViewingState.tsx
@@ -8,7 +8,7 @@ import {
   Typography,
 } from "@material-ui/core";
 import React, { FunctionComponent } from "react";
-import { WritingStyle } from "../../utils/mcaWritingStyleFormatter";
+import { WritingStyle } from "../../utils/writingStyle";
 import { FieldValue, FieldValueTypes } from "./FieldValue";
 
 type IFieldValue = string | undefined;
