--- src/utils/dateTimeConverter.tsx
@@ -1,1 +1,0 @@
-export const isoDate = (isoDateTime: string) => isoDateTime.slice(0, 10);
