+++ src/utils/writingStyle.test.tsx
@@ -1,45 +1,12 @@
 import { FieldValueTypes } from "../components/dataPanel/FieldValue";
 import { Activities, Environments, Purposes } from "../entities/IUse";
 import {
-  formatDateLong,
-  formatDateShort,
   formatFieldValue,
   formatOwners,
   formatUses,
   Placeholders,
   titleCase,
-} from "./mcaWritingStyleFormatter";
-
-describe("formatDateLong()", () => {
-  const expectations = [
-    { in: "1 April 2021", out: "1 Apr 21" },
-    { in: "1 April 2022", out: "1 Apr 22" },
-    { in: "31 October 2028", out: "31 Oct 28" },
-  ];
-
-  expectations.forEach((expectation) => {
-    it(`formats ${JSON.stringify(expectation.in)} ==> ${
-      expectation.out
-    }`, () => {
-      expect(formatDateLong(expectation.in)).toEqual(expectation.out);
-    });
-  });
-});
-
-describe("formatDateShort()", () => {
-  const expectations = [
-    { in: "2020-02-01T00:00:00.000Z", out: "Feb 2020" },
-    { in: "2021-05-06T10:00:03.592854", out: "May 2021" },
-  ];
-
-  expectations.forEach((expectation) => {
-    it(`formats ${JSON.stringify(expectation.in)} ==> ${
-      expectation.out
-    }`, () => {
-      expect(formatDateShort(expectation.in)).toEqual(expectation.out);
-    });
-  });
-});
+} from "./writingStyle";
 
 describe("titleCase()", () => {
   const expectations = [
